backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq
torch==2.1.0
transformers==4.35.0
pydub==0.25.1
reportlab==4.0.7
numpy==1.24.3
supabase
python-dotenv
